+++ scripts/installer/local-installer.js
@@ -303,3 +303,6 @@ node bot.js
 
 module.exports = LocalInstaller;
 
+
+
+